Basic front-end done (barring one issue)
I have completed the implementation of all of the basic front-end components (with empty div boxes reflecting areas that images will go).
There is a slight quirk with the "remove exercise" button; when it gets clicked, every exercise below that button also disappears.
I tried to fix this for a while but did not succeed. Other than that, everything works nicely (although we can adjust the aesthetics of it if needed).